Wrap kubeconfig flag handling in restconfig.Producer
Moving the execution function into restconfig.Producer will ensure
that flags can be handled entirely inside this package, reducing the
exposure of internals to subctl commands, and the knowledge required
to implement commands.

For now only single-context commands are handled, multi-context
commands will come later. configAndOverrides will be useful then.

This also removes ForClusters() and Clients() which are effectively
unused, and unexports ClusterNameFromContext() which is only used in
the restconfig package.

Signed-off-by: Stephen Kitt <skitt@redhat.com>